SignInCallbacksController の修正
