+++ .github/workflows/build.yml
@@ -88,6 +88,7 @@ jobs:
           ./cc-test-reporter before-build
           bundle exec rspec \
             spec/requests/api/internal/sign_in/create_spec.rb \
+            spec/requests/sign_in_callbacks/show_spec.rb \
             spec/requests/sign_in/new_spec.rb \
             spec/requests/welcome/show_spec.rb
           ./cc-test-reporter after-build --coverage-input-type simplecov --exit-code $?
