+++ config/routes/web.rb
@@ -158,6 +158,7 @@ match "/@:username/records/:record_id", via: :patch, as: :record, to: "records#u
 match "/episodes/:episode_id/records", via: :post, as: :episode_record_list, to: "episode_records#create"
 match "/sign_in", via: :get, as: :new_user_session, to: "sign_in#new" # for Devise
 match "/sign_in", via: :get, as: :sign_in, to: "sign_in#new"
+match "/sign_in/callback", via: :get, as: :sign_in_callback, to: "sign_in_callbacks#show"
 match "/track", via: :get, as: :track, to: "tracks#show"
 match "/works/:anime_id/episodes/:episode_id", via: :get, as: :episode, to: "episodes#show"
 
@@ -166,7 +167,6 @@ scope module: :v4 do
     devise_scope :user do
       match "/registrations", via: :post, as: :registrations, to: "registrations#create"
       match "/registrations/new", via: :get, as: :new_registration, to: "registrations#new"
-      match "/sign_in/callback", via: :get, as: :sign_in_callback, to: "sign_in_callbacks#show"
       match "/sign_up", via: :get, as: :sign_up, to: "sign_up#new"
       match "/sign_up", via: :post, to: "sign_up#create"
       match "/user_email", via: :patch, as: :user_email, to: "user_emails#update"
