+++ spec/requests/sign_in_callbacks/show_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 describe "GET /sign_in/callback", type: :request do
-  context "when request is valid" do
+  context "パラメータが正常なとき" do
     let(:user) { create :registered_user }
     let!(:email_confirmation) { create(:email_confirmation, email: user.email, event: "sign_in", back: "/home") }
 
-    it "signs in" do
+    it "ログインできること" do
       expect(EmailConfirmation.count).to eq 1
 
       get "/sign_in/callback", params: {token: email_confirmation.token}
